+++ src/constants/facts-meta-keys.ts
@@ -3,9 +3,11 @@ export enum FactsMetaKeys {
   RETRIES_LIMIT = 'retriesLimit',
   EXECUTE_AFTER = 'executeAfter',
   EXPIRE_AFTER = 'expireAfter',
-  COMPENSATOR_NODE = 'compensatorNode',
+  ROLLBACK_WHEN_ERROR_NODE = 'rollbackWhenErrorNode',
+  ROLLBACK_WHEN_SUCCESS_NODE = 'rollbackWhenSuccessNode',
   TIMEOUT_BETWEEN_RETRIES = 'timeoutBetweenRetries',
   LAST_RETRY_TIME = 'lastRetryTime',
+  RUN_AFTER_NODES_SUCCEED = 'runAfterNodesSucceed',
 }
 
 export const factsMetaKeys: string[] = Object.values(FactsMetaKeys);