+++ src/framework/framework-interface.ts
@@ -1,7 +1,7 @@
 import { Facts } from '@parameters';
 
 export interface FrameworkInterface<DataType, NodeName extends string> {
-  next(node: NodeName, facts: Facts<DataType, NodeName>): void;
+  next(nodes: NodeName | NodeName[], facts: Facts<DataType, NodeName>): void;
   exit(facts: Facts<DataType, NodeName>, error?: Error): void;
   retry(node: NodeName, facts: Facts<DataType, NodeName>, error?: Error): void;
 }