+++ src/options/framework-options.ts
@@ -1,12 +1,12 @@
 import { EventEmitter } from 'events';
 import { Logger } from '@logger';
-import { FactsMeta } from '@parameters';
+import { NodeMeta } from '@parameters';
 import { Processor } from '@node';
 
 export type FrameworkOptions<DataType, NodeName extends string> = {
   eventEmitter: EventEmitter;
   logger: Logger;
-  meta: Map<NodeName, FactsMeta>;
+  meta: Map<NodeName, NodeMeta>;
   nodes: Map<NodeName, Processor<DataType, NodeName>>;
   verbose: boolean;
 };