+++ src/parameters/index.ts
@@ -1,4 +1,5 @@
 export * from './facts';
 export * from './facts-meta';
+export * from './node-meta';
 export * from './executor';
 export * from './pseudo-interval-params';