+++ src/parameters/node-meta.ts
@@ -1,0 +1,8 @@
+export type NodeMeta = {
+  retriesLimit: number;
+  lastRetryTime: number;
+  timeoutBetweenRetries: number;
+  rollbackWhenErrorNode?: string;
+  rollbackWhenSuccessNode?: string;
+  runAfterNodesSucceed?: string[];
+};